requirements.txt
ipython 
pillow==9.5.0
cherrypy
pytest